Fix For SKR E3 Mini V2.0 Parts Cooling Fan Bug
This fix may only be relevant to the board listed in the title, for which I would not now how to modify for it specifically.
After flashing from current Marlin (2021-02-13 Bugfix) the parts cooling fan on this board stopped functioning correctly, commands to the fan only enacted after (any) menu operation occurred (tested BTT TFT35 v3.0 emulation mode) then the fan began performing the current state. I searched the web and found a reddit thread on this issue that ultimately led to a user reporting the problem solved updating to the latest binary from the BTT repo. I checked the latest set of changes merged and found a commit reference to a fan bug, searched the commit to find the changes included there that could be related to the fan (only in the platformio.ini file).
I rebuilt with these additions, flashed, and the fan now functions properly again.

Source: https://github.com/bigtreetech/BIGTREETECH-SKR-mini-E3/commit/1fc9de352816c3961fcf0bb2799795832e5ceba4